requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-community>=0.0.10
python-dotenv>=1.0.0
pydantic>=2.0.0
sympy>=1.12
httpx>=0.26.0
requests>=2.31.0